Add @Before and @After methods to AbstractTest class; remove redundant code from tests
