+++ src/test/java/org/diplom3/HeaderTabsTest.java
@@ -1,0 +1,4 @@
+package org.diplom3;
+
+public class HeaderTabsTest extends AbstractTest {
+}
